+++ packages/cursorless-engine/src/processTargets/targets/PlainTarget.ts
@@ -1,3 +1,5 @@
+import {tryConstructTarget} from "../../util/tryConstructTarget";
+import {TextEditor, Range} from "@cursorless/common";
 import { BaseTarget, CommonTargetParameters } from "./BaseTarget";
 
 interface PlainTargetParameters extends CommonTargetParameters {
@@ -32,3 +34,21 @@ export class PlainTarget extends BaseTarget<PlainTargetParameters> {
     };
   }
 }
+
+/**
+ * Constructs a {@link PlainTarget} from the given range, or returns undefined
+ * if the range is undefined
+ * @param editor The editor containing the range
+ * @param range The range to convert into a target
+ * @param isReversed Whether the rain should be backward
+ * @returns A new {@link PlainTarget} constructed from the given range, or null
+ * if the range is undefined
+ */
+export function tryConstructPlainTarget(
+  editor: TextEditor,
+  range: Range | undefined,
+  isReversed: boolean
+): PlainTarget | undefined {
+  return tryConstructTarget(PlainTarget, editor, range, isReversed);
+}
+
