+++ packages/cursorless-engine/src/util/tryConstructTarget.ts
@@ -1,8 +1,6 @@
 import { Range, TextEditor } from "@cursorless/common";
 import {
   CommonTargetParameters,
-  LineTarget,
-  PlainTarget,
 } from "../processTargets/targets";
 import { Target } from "../typings/target.types";
 
@@ -33,38 +31,4 @@ export function tryConstructTarget<T extends Target>(
         isReversed,
         contentRange: range,
       });
-}
-
-/**
- * Constructs a {@link PlainTarget} from the given range, or returns undefined
- * if the range is undefined
- * @param editor The editor containing the range
- * @param range The range to convert into a target
- * @param isReversed Whether the rain should be backward
- * @returns A new {@link PlainTarget} constructed from the given range, or null
- * if the range is undefined
- */
-export function tryConstructPlainTarget(
-  editor: TextEditor,
-  range: Range | undefined,
-  isReversed: boolean,
-): PlainTarget | undefined {
-  return tryConstructTarget(PlainTarget, editor, range, isReversed);
-}
-
-/**
- * Constructs a {@link LineTarget} from the given range, or returns undefined
- * if the range is undefined
- * @param editor The editor containing the range
- * @param range The range to convert into a target
- * @param isReversed Whether the rain should be backward
- * @returns A new {@link LineTarget} constructed from the given range, or null
- * if the range is undefined
- */
-export function constructLineTarget(
-  editor: TextEditor,
-  range: Range | undefined,
-  isReversed: boolean,
-): LineTarget | undefined {
-  return tryConstructTarget(LineTarget, editor, range, isReversed);
 }
